+++ src/event/mod.rs
@@ -1,2 +1,3 @@
+pub mod correlation;
 pub mod payload;
 pub mod time;